Relocate utils package
